src: patch_track: Adiciona nova funcao register_patch_track
To support the new feature `kw patch-track`, it should be able to
register the patches that are sent with kw send_patches. In this sense,
add a new function register-patch-track that is called after the send_patch
mail_send function, registering infos from the patch as an id, the status,
date, time and title.

Signed-off-by: JoaoGBSouza <joao.souza@upp.com.br>